requirements.txt
pyjwt
requests
six